backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0.post1

# CSV uploads, analytics, and reporting
pandas==2.1.1
reportlab==4.0.7
openpyxl==3.1.2

# Serving static files
aiofiles==23.2.1

# ML predictions
scikit-learn==1.3.2
joblib==1.3.2

# Security enhancements
python-jose[cryptography]==3.3.0
slowapi==0.1.8
gunicorn==21.2.0

# Production enhancements
prometheus-client==0.17.1  # For metrics
psutil==5.9.6  # For system metrics in health checks
uvloop==0.19.0  # For faster event loop on Linux/macOS
httptools==0.6.1  # For faster HTTP parsing
pyjwt==2.8.0  # For JWT handling
starlette-exporter==0.17.1  # For Prometheus metrics
loguru==0.7.2  # For better logging
asgi-correlation-id==4.2.0  # For request tracing

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1

# Code quality
ruff==0.1.4
black==23.10.1

asgi-correlation-id